Add rotation and scaling to the objects.
Organise the startup more with promises and creating a dedicated
index.js to more easily order the startup process. This frees up
scene.js to be specifically about rending the scene objects.

Closes #39.
