+++ static/scripts/applicationDataExample.js
@@ -1,0 +1,34 @@
+export var applicationData = {
+    applications: [
+        {
+            name: "app1",
+            color: "#0287fc",
+            servers: [
+                {
+                    name: "app1hostname2"
+                },
+                {
+                    name: "app1hostname2"
+                }
+            ],
+            position: [0, 0, 0],
+            rotation: [0, 0, 0],
+            scale: [2, 1, 1]
+        },
+        {
+            name: "app2",
+            color: "#06f7fc",
+            servers: [
+                {
+                    name: "app2hostname1"
+                },
+                {
+                    name: "app2hostname2"
+                }
+            ],
+            position: [10, 0, 0],
+            rotation: [30, 30, 0],
+            scale: [1, 1, 1]
+        }
+    ]
+}
